월간막대카드 폰트사이즈 증가, res type적용
