+++ client/web/src/organisms/mainpage/ranking/types/ScoresBarChart.types.ts
@@ -1,0 +1,8 @@
+import { MonthlyScoresItem } from '@truepoint/shared/dist/res/RankingsResTypes.interface';
+
+export interface ScoresBarChartProps{
+  data: MonthlyScoresItem[],
+  loading?: boolean,
+  column? : string,
+  barColor? : string,
+}
